server/requirements.txt
aniso8601==9.0.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
marshmallow==3.14.1
marshmallow-sqlalchemy==0.28.0
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3
